Windows: Update legacy app.vcproj for wxWidgets 3
.gitignore
README_WIN.txt
app.vcproj

- Ignore .ncb files.

- Change lib prefix from wxmsw26 to wxmsw31u.
The u is now necessary for UNICODE builds.

- Change dir prefix from C:\wxWidgets-2.6.3\ to a more generic
C:\wxWidgets\.

- Enable Run Time Type Info (RTTI).
By default wxWidgets is built with RTTI and there cannot be an RTTI
mismatch between the app and the static libraries it links to.
http://trac.wxwidgets.org/ticket/16431

- Create GalaXQL.dat as a post-build event.

- Change SQLite3 reference from individual source files to amalgamation.

- Remove reference to non-existent icon1.ico.

- Remove libc.lib from being ignored as a default library.
Ignoring libc isn't necessary; if it is then something's wrong and that
should be brought to the developer's attention, not ignored.

- Add a legacy warning (README_WIN.txt) and reference it in the project.